+++ user/include/rainbow/arch/ia32/syscall.h
@@ -40,9 +40,9 @@ extern "C" {
 #define __SYSENTER "call *0xEFFFF000\n"
 
 
-static inline long __syscall0(long function)
+static inline intptr_t __syscall0(intptr_t function)
 {
-    long result;
+    intptr_t result;
 
     asm volatile (
         __SYSENTER
@@ -55,9 +55,9 @@ static inline long __syscall0(long function)
 }
 
 
-static inline long __syscall1(long function, long arg1)
+static inline intptr_t __syscall1(intptr_t function, intptr_t arg1)
 {
-    long result;
+    intptr_t result;
 
     asm volatile (
         __SYSENTER
@@ -71,9 +71,9 @@ static inline long __syscall1(long function, long arg1)
 }
 
 
-static inline long __syscall2(long function, long arg1, long arg2)
+static inline intptr_t __syscall2(intptr_t function, intptr_t arg1, intptr_t arg2)
 {
-    long result;
+    intptr_t result;
 
     asm volatile (
         __SYSENTER
@@ -88,9 +88,9 @@ static inline long __syscall2(long function, long arg1, long arg2)
 }
 
 
-static inline long __syscall3(long function, long arg1, long arg2, long arg3)
+static inline intptr_t __syscall3(intptr_t function, intptr_t arg1, intptr_t arg2, intptr_t arg3)
 {
-    long result;
+    intptr_t result;
 
     asm volatile (
         __SYSENTER
@@ -106,9 +106,9 @@ static inline long __syscall3(long function, long arg1, long arg2, long arg3)
 }
 
 
-static inline long __syscall4(long function, long arg1, long arg2, long arg3, long arg4)
+static inline intptr_t __syscall4(intptr_t function, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4)
 {
-    long result;
+    intptr_t result;
 
     asm volatile (
         __SYSENTER
@@ -125,9 +125,9 @@ static inline long __syscall4(long function, long arg1, long arg2, long arg3, lo
 }
 
 
-static inline long __syscall5(long function, long arg1, long arg2, long arg3, long arg4, long arg5)
+static inline intptr_t __syscall5(intptr_t function, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5)
 {
-    long result;
+    intptr_t result;
 
     asm volatile (
         __SYSENTER
@@ -150,7 +150,7 @@ static inline long __syscall5(long function, long arg1, long arg2, long arg3, lo
     so for now, we will have libc implement it as a non-inline function.
 */
 
-long __syscall6(long function, long arg1, long arg2, long arg3, long arg4, long arg5, long arg6);
+intptr_t __syscall6(intptr_t function, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6);
 
 
 #undef __SYSENTER
